Silence the compiler in 64 bit architectures.
Using the c99 %zu and misc casts.

+++ pn_oim.c
@@ -206,7 +206,7 @@ send_receive_request (PnNode *conn,
     {
         gsize len;
         pn_node_write (conn, header, strlen (header), &len, NULL);
-        pn_debug ("write_len=%d", len);
+        pn_debug ("write_len=%zu", len);
     }
 
     g_free (header);
@@ -268,7 +268,7 @@ send_delete_request (PnNode *conn,
     {
         gsize len;
         pn_node_write (conn, header, strlen (header), &len, NULL);
-        pn_debug ("write_len=%d", len);
+        pn_debug ("write_len=%zu", len);
     }
 
     g_free (header);
@@ -387,7 +387,7 @@ send_send_request (PnNode *conn,
     {
         gsize len;
         pn_node_write (conn, header, strlen (header), &len, NULL);
-        pn_debug ("write_len=%d", len);
+        pn_debug ("write_len=%zu", len);
     }
 
     g_free (header);
@@ -482,7 +482,7 @@ send_auth_request (PnNode *conn,
     {
         gsize len;
         pn_node_write (conn, header, strlen (header), &len, NULL);
-        pn_debug ("write_len=%d", len);
+        pn_debug ("write_len=%zu", len);
     }
 
     g_free (header);
@@ -555,7 +555,7 @@ process_body_receive (OimRequest *oim_request,
     gchar *cur;
     guint32 date = 0;
 
-    pn_debug("body=[%.*s]", length, body);
+    pn_debug("body=[%.*s]", (guint) length, body);
 
     /** @todo find a way to parse the date in win32 */
 #ifndef G_OS_WIN32
@@ -600,7 +600,7 @@ process_body_delete (OimRequest *oim_request,
                      char *body,
                      gsize length)
 {
-    pn_debug("body=[%.*s]", length, body);
+    pn_debug("body=[%.*s]", (guint) length, body);
 
     if (strstr (body, "Schema validation error"))
         pn_error ("deleting oim=[%s]: schema validation error", oim_request->message_id);
@@ -613,7 +613,7 @@ process_body_send (OimRequest *oim_request,
 {
     gchar *cur;
 
-    pn_debug("body=[%.*s]", length, body);
+    pn_debug("body=[%.*s]", (guint) length, body);
 
     cur = strstr (body, "<LockKeyChallenge ");
     if (cur)
@@ -675,7 +675,7 @@ process_body_auth (OimRequest *oim_request,
 {
     gchar *cur;
 
-    pn_debug ("body=[%.*s]", length, body);
+    pn_debug ("body=[%.*s]", (guint) length, body);
 
     cur = strstr (body, "<wsse:BinarySecurityToken Id=\"PPToken1\">");
     if (cur)
